Fix account.route.spec.
Move server's web3 initialization into its own JS module.
Fix some values in .env.example (remove single quotes).
