+++ server/.env.example
@@ -4,5 +4,5 @@ PLAID_PUBLIC_KEY=1234567890abcdef1234567890abcd
 PLAID_ENV=sandbox
 PRIVATE_KEY=0x3daa79a26454a5528a3523f9e6345efdbd636e63f8c24a835204e6ccb5c88000
 POBA_ERC725_CONTRACT_ADDRESS=0xd591742c78ab7cf44761d6126dcd26e0a869ed883c41fc3cf46c660c6a5c8df1
-POBA_ERC725_URI='http://poba.poa.network'
+POBA_ERC725_URI=http://poba.poa.network
 RPC=http://localhost:8545
